+++ tests/Unit/Handler/PropertyMapperTest.php
@@ -7,7 +7,6 @@ namespace JsonMapper\Tests\Unit\Handler;
 use JsonMapper\Builders\PropertyBuilder;
 use JsonMapper\Cache\NullCache;
 use JsonMapper\Enums\Visibility;
-use JsonMapper\Exception\TypeError;
 use JsonMapper\Handler\FactoryRegistry;
 use JsonMapper\Handler\PropertyMapper;
 use JsonMapper\JsonMapperFactory;
@@ -25,6 +24,7 @@ use JsonMapper\Tests\Implementation\Popo;
 use JsonMapper\Tests\Implementation\PrivatePropertyWithoutSetter;
 use JsonMapper\Tests\Implementation\SimpleObject;
 use JsonMapper\Tests\Implementation\UserWithConstructorParent;
+use JsonMapper\ValueObjects\ArrayInformation;
 use JsonMapper\ValueObjects\PropertyMap;
 use JsonMapper\Wrapper\ObjectWrapper;
 use PHPUnit\Framework\TestCase;
@@ -55,7 +55,7 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('value')
-            ->addType($type, false)
+            ->addType($type, ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -78,7 +78,7 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('createdAt')
-            ->addType(\DateTimeImmutable::class, false)
+            ->addType(\DateTimeImmutable::class, ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -102,7 +102,7 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('child')
-            ->addType(SimpleObject::class, false)
+            ->addType(SimpleObject::class, ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PRIVATE())
             ->build();
@@ -134,7 +134,7 @@ class PropertyMapperTest extends TestCase
     {
         $fileProperty = PropertyBuilder::new()
             ->setName('ids')
-            ->addType('int', true)
+            ->addType('int', ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -153,11 +153,35 @@ class PropertyMapperTest extends TestCase
     /**
      * @covers \JsonMapper\Handler\PropertyMapper
      */
+    public function testPublicScalarValueMultiDimensionalArrayIsSet(): void
+    {
+        $fileProperty = PropertyBuilder::new()
+            ->setName('ids')
+            ->addType('int', ArrayInformation::multiDimension(2))
+            ->setIsNullable(false)
+            ->setVisibility(Visibility::PUBLIC())
+            ->build();
+        $propertyMap = new PropertyMap();
+        $propertyMap->addProperty($fileProperty);
+        $value = [1 => [2, 3], 2 => [2, 3], 3 => [2, 3]];
+        $json = (object) ['ids' => $value];
+        $object = new \stdClass();
+        $wrapped = new ObjectWrapper($object);
+        $propertyMapper = new PropertyMapper();
+
+        $propertyMapper->__invoke($json, $wrapped, $propertyMap, $this->createMock(JsonMapperInterface::class));
+
+        self::assertEquals($value, $object->ids);
+    }
+
+    /**
+     * @covers \JsonMapper\Handler\PropertyMapper
+     */
     public function testPublicCustomClassArrayIsSet(): void
     {
         $property = PropertyBuilder::new()
             ->setName('children')
-            ->addType(SimpleObject::class, true)
+            ->addType(SimpleObject::class, ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PRIVATE())
             ->build();
@@ -183,11 +207,57 @@ class PropertyMapperTest extends TestCase
     /**
      * @covers \JsonMapper\Handler\PropertyMapper
      */
+    public function testPublicCustomClassMultidimensionalArrayIsSet(): void
+    {
+        $property = PropertyBuilder::new()
+            ->setName('children')
+            ->addType(SimpleObject::class, ArrayInformation::multiDimension(2))
+            ->setIsNullable(false)
+            ->setVisibility(Visibility::PUBLIC())
+            ->build();
+        $propertyMap = new PropertyMap();
+        $propertyMap->addProperty($property);
+        $jsonMapper = $this->createMock(JsonMapperInterface::class);
+        $jsonMapper->expects(self::exactly(4))
+            ->method('mapObject')
+            ->with(self::isInstanceOf(\stdClass::class), self::isInstanceOf(SimpleObject::class))
+            ->willReturnCallback(static function ($json, $object) {
+                $object->setName($json->name);
+            });
+        $json = (object) ['children' => [
+            [
+                (object) ['name' => __NAMESPACE__],
+                (object) ['name' => __FUNCTION__]
+            ],
+            [
+                (object) ['name' => __NAMESPACE__],
+                (object) ['name' => __FUNCTION__]
+            ],
+        ]];
+        $object = new \stdClass();
+        $wrapped = new ObjectWrapper($object);
+        $propertyMapper = new PropertyMapper();
+
+        $propertyMapper->__invoke($json, $wrapped, $propertyMap, $jsonMapper);
+
+        self::assertCount(2, $object->children);
+        self::assertEquals(
+            [
+                [new SimpleObject(__NAMESPACE__), new SimpleObject(__FUNCTION__)],
+                [new SimpleObject(__NAMESPACE__), new SimpleObject(__FUNCTION__)],
+            ],
+            $object->children
+        );
+    }
+
+    /**
+     * @covers \JsonMapper\Handler\PropertyMapper
+     */
     public function testArrayPropertyIsCasted(): void
     {
         $property = PropertyBuilder::new()
             ->setName('notes')
-            ->addType('string', true)
+            ->addType('string', ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -211,7 +281,7 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('user')
-            ->addType(UserWithConstructor::class, false)
+            ->addType(UserWithConstructor::class, ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -242,7 +312,7 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('user')
-            ->addType(UserWithConstructor::class, true)
+            ->addType(UserWithConstructor::class, ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -275,12 +345,64 @@ class PropertyMapperTest extends TestCase
     /**
      * @covers \JsonMapper\Handler\PropertyMapper
      */
+    public function testCanMapPropertyAsMultiDimensionalArrayWithClassFactory(): void
+    {
+        $property = PropertyBuilder::new()
+            ->setName('userHistory')
+            ->addType(UserWithConstructor::class, ArrayInformation::multiDimension(2))
+            ->setIsNullable(false)
+            ->setVisibility(Visibility::PUBLIC())
+            ->build();
+        $propertyMap = new PropertyMap();
+        $propertyMap->addProperty($property);
+        $jsonMapper = $this->createMock(JsonMapperInterface::class);
+        $json = (object) ['userHistory' => [
+            '2021-02-03' => [
+                'original' => (object) ['id' => 1234, 'name' => 'John Doe'],
+                'new' => (object) ['id' => 1234, 'name' => 'Johnathan Doe'],
+            ],
+            '2022-08-16' => [
+                'original' => (object) ['id' => 1234, 'name' => 'Johnathan Doe'],
+                'new' => (object) ['id' => 1234, 'name' => 'J. Doe'],
+            ],
+        ]];
+        $object = new UserWithConstructorParent();
+        $wrapped = new ObjectWrapper($object);
+        $classFactoryRegistry = FactoryRegistry::withNativePhpClassesAdded();
+        $classFactoryRegistry->addFactory(
+            UserWithConstructor::class,
+            static function ($params) {
+                return new UserWithConstructor($params->id, $params->name);
+            }
+        );
+        $propertyMapper = new PropertyMapper($classFactoryRegistry);
+
+        $propertyMapper->__invoke($json, $wrapped, $propertyMap, $jsonMapper);
+
+        self::assertEquals(
+            [
+                '2021-02-03' => [
+                    'original' => new UserWithConstructor(1234, 'John Doe'),
+                    'new' => new UserWithConstructor(1234, 'Johnathan Doe'),
+                ],
+                '2022-08-16' => [
+                    'original' => new UserWithConstructor(1234, 'Johnathan Doe'),
+                    'new' => new UserWithConstructor(1234, 'J. Doe'),
+                ],
+            ],
+            $object->userHistory
+        );
+    }
+
+    /**
+     * @covers \JsonMapper\Handler\PropertyMapper
+     */
     public function testCanMapUnionPropertyAsArrayWithClassFactory(): void
     {
         $property = PropertyBuilder::new()
             ->setName('user')
-            ->addType(UserWithConstructor::class, true)
-            ->addType(\DateTime::class, true)
+            ->addType(UserWithConstructor::class, ArrayInformation::singleDimension())
+            ->addType(\DateTime::class, ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -317,7 +439,7 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('child')
-            ->addType(SimpleObject::class, false)
+            ->addType(SimpleObject::class, ArrayInformation::notAnArray())
             ->setIsNullable(true)
             ->setVisibility(Visibility::PRIVATE())
             ->build();
@@ -342,7 +464,7 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('child')
-            ->addType(SimpleObject::class, false)
+            ->addType(SimpleObject::class, ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PRIVATE())
             ->build();
@@ -368,7 +490,7 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('number')
-            ->addType('int', false)
+            ->addType('int', ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PRIVATE())
             ->build();
@@ -396,11 +518,11 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('value')
-            ->addType('int', false)
-            ->addType('double', false)
-            ->addType('float', false)
-            ->addType('string', false)
-            ->addType('bool', false)
+            ->addType('int', ArrayInformation::notAnArray())
+            ->addType('double', ArrayInformation::notAnArray())
+            ->addType('float', ArrayInformation::notAnArray())
+            ->addType('string', ArrayInformation::notAnArray())
+            ->addType('bool', ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -426,10 +548,10 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('values')
-            ->addType('int', true)
-            ->addType('float', true)
-            ->addType('string', true)
-            ->addType('bool', true)
+            ->addType('int', ArrayInformation::singleDimension())
+            ->addType('float', ArrayInformation::singleDimension())
+            ->addType('string', ArrayInformation::singleDimension())
+            ->addType('bool', ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -454,8 +576,8 @@ class PropertyMapperTest extends TestCase
         $now = new \DateTime();
         $property = PropertyBuilder::new()
             ->setName('moment')
-            ->addType('int', true)
-            ->addType(\DateTime::class, true)
+            ->addType('int', ArrayInformation::singleDimension())
+            ->addType(\DateTime::class, ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -479,8 +601,8 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('user')
-            ->addType(User::class, false)
-            ->addType(Popo::class, false)
+            ->addType(User::class, ArrayInformation::notAnArray())
+            ->addType(Popo::class, ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -505,8 +627,8 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('users')
-            ->addType(User::class, true)
-            ->addType(Popo::class, true)
+            ->addType(User::class, ArrayInformation::singleDimension())
+            ->addType(Popo::class, ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();
@@ -596,7 +718,7 @@ class PropertyMapperTest extends TestCase
             ->setName('shape')
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
-            ->addType($type, false)
+            ->addType($type, ArrayInformation::notAnArray())
             ->build();
         $propertyMap = new PropertyMap();
         $propertyMap->addProperty($property);
@@ -622,7 +744,7 @@ class PropertyMapperTest extends TestCase
             ->setName('shape')
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
-            ->addType($type, false)
+            ->addType($type, ArrayInformation::notAnArray())
             ->build();
         $propertyMap = new PropertyMap();
         $propertyMap->addProperty($property);
@@ -650,7 +772,7 @@ class PropertyMapperTest extends TestCase
             ->setName('status')
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
-            ->addType(\JsonMapper\Tests\Implementation\Php81\Status::class, false)
+            ->addType(\JsonMapper\Tests\Implementation\Php81\Status::class, ArrayInformation::notAnArray())
             ->build();
         $propertyMap = new PropertyMap();
         $propertyMap->addProperty($property);
@@ -678,7 +800,7 @@ class PropertyMapperTest extends TestCase
             ->setName('historicStates')
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
-            ->addType(\JsonMapper\Tests\Implementation\Php81\Status::class, true)
+            ->addType(\JsonMapper\Tests\Implementation\Php81\Status::class, ArrayInformation::singleDimension())
             ->build();
         $propertyMap = new PropertyMap();
         $propertyMap->addProperty($property);
@@ -698,12 +820,51 @@ class PropertyMapperTest extends TestCase
 
     /**
      * @covers \JsonMapper\Handler\PropertyMapper
+     * @requires PHP >= 8.1
+     */
+    public function testItCanMapEnumMultiDimensionalArrayType(): void
+    {
+        $json = (object) ['historicStatesByDate' => [
+            '2022-08-16' => ['draft', 'published'],
+            '2022-08-22' => ['archived']
+        ]];
+        $object = new \JsonMapper\Tests\Implementation\Php81\BlogPost();
+        $wrapped = new ObjectWrapper($object);
+        $property = PropertyBuilder::new()
+            ->setName('historicStatesByDate')
+            ->setIsNullable(false)
+            ->setVisibility(Visibility::PUBLIC())
+            ->addType(\JsonMapper\Tests\Implementation\Php81\Status::class, ArrayInformation::multiDimension(2))
+            ->build();
+        $propertyMap = new PropertyMap();
+        $propertyMap->addProperty($property);
+        $expected = new BlogPost();
+        $expected->historicStatesByDate = [
+            '2022-08-16' => [
+                \JsonMapper\Tests\Implementation\Php81\Status::DRAFT,
+                \JsonMapper\Tests\Implementation\Php81\Status::PUBLISHED,
+            ],
+            '2022-08-22' => [
+                \JsonMapper\Tests\Implementation\Php81\Status::ARCHIVED,
+            ]
+        ];
+
+        $propertyMapper = new PropertyMapper();
+        $jsonMapper = (new JsonMapperFactory())->create($propertyMapper, new DocBlockAnnotations(new NullCache()));
+
+        $propertyMapper->__invoke($json, $wrapped, $propertyMap, $jsonMapper);
+
+        self::assertEquals($expected, $object);
+    }
+
+    /**
+     * @covers \JsonMapper\Handler\PropertyMapper
      */
     public function testItThrowsExceptionForNonExistingClass(): void
     {
         $property = PropertyBuilder::new()
             ->setName('child')
-            ->addType("\Some\Non\Existing\Class", false)
+            ->addType("\Some\Non\Existing\Class", ArrayInformation::notAnArray())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PRIVATE())
             ->build();
@@ -727,8 +888,8 @@ class PropertyMapperTest extends TestCase
     {
         $property = PropertyBuilder::new()
             ->setName('value')
-            ->addType('string', false)
-            ->addType('string', true)
+            ->addType('string', ArrayInformation::notAnArray())
+            ->addType('string', ArrayInformation::singleDimension())
             ->setIsNullable(false)
             ->setVisibility(Visibility::PUBLIC())
             ->build();