+++ tests/Unit/ValueObjects/PropertyTypeTest.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace JsonMapper\Tests\ValueObjects\Unit;
+namespace JsonMapper\Tests\Unit\ValueObjects;
 
+use JsonMapper\ValueObjects\ArrayInformation;
 use JsonMapper\ValueObjects\PropertyType;
 use PHPUnit\Framework\TestCase;
 
@@ -14,10 +15,11 @@ class PropertyTypeTest extends TestCase
      */
     public function testGettersReturnConstructorValues(): void
     {
-        $propertyType = new PropertyType('int', false);
+        $arrayInformation = ArrayInformation::notAnArray();
+        $propertyType = new PropertyType('int', $arrayInformation);
 
         self::assertSame('int', $propertyType->getType());
-        self::assertFalse($propertyType->isArray());
+        self::assertEquals($arrayInformation, $propertyType->getArrayInformation());
     }
 
     /**
@@ -25,14 +27,54 @@ class PropertyTypeTest extends TestCase
      */
     public function testCanBeConvertedToJson(): void
     {
-        $propertyType = new PropertyType('int', false);
+        $propertyType = new PropertyType('int', ArrayInformation::notAnArray());
 
         $propertyAsJson = json_encode($propertyType);
 
         self::assertIsString($propertyAsJson);
         self::assertJsonStringEqualsJsonString(
-            '{"type":"int","isArray":false}',
+            '{"type":"int","isArray":false,"arrayInformation":{"isArray":false,"dimensions":0}}',
             (string) $propertyAsJson
         );
+    }
+
+    /**
+     * @covers \JsonMapper\ValueObjects\PropertyType
+     * @dataProvider isArrayValueAndExpectation
+     */
+    public function testIsArrayReturnsCorrectForPossibleValues(ArrayInformation $isArray, bool $expected): void
+    {
+        $propertyType = new PropertyType('int', $isArray);
+
+        self::assertSame($expected, $propertyType->isArray());
+    }
+
+    /**
+     * @covers \JsonMapper\ValueObjects\PropertyType
+     * @dataProvider isMultiDimensionalArrayValueAndExpectation
+     */
+    public function testIsMultiDimensionalArrayReturnsCorrectForPossibleValues(ArrayInformation $isArray, bool $expected): void
+    {
+        $propertyType = new PropertyType('int', $isArray);
+
+        self::assertSame($expected, $propertyType->isMultiDimensionalArray());
+    }
+
+    public function isArrayValueAndExpectation(): array
+    {
+        return [
+            'no' => [ArrayInformation::notAnArray(), false],
+            'single dimensional' => [ArrayInformation::singleDimension(), true],
+            'multi dimensional' => [ArrayInformation::multiDimension(2), true,]
+        ];
+    }
+
+    public function isMultiDimensionalArrayValueAndExpectation(): array
+    {
+        return [
+            'no' => [ArrayInformation::notAnArray(), false],
+            'single dimensional' => [ArrayInformation::singleDimension(), false],
+            'multi dimensional' => [ArrayInformation::multiDimension(2), true,]
+        ];
     }
 }